fix apiService home
